feat: I added the header file and demo folder
This header file has the ind2sub2D function only, which converts indexes
to bidimensional subscripts. All indexes and subscripts start from 0.

- The "arrayOfDimensions" is an input that holds the number of
elements in the two dimensions of this array.
- The "index" input represents the linear index to be converted.
- And the "convertedIndex" is an output array containing both subscripts.
The first is the row, and the second is the column.

See the demo/demo.c file for more details.