feat: Add the multidimensional case for sub2ind
The new sub2ind function is for cases of order two or higher. It
calculates all the indices iteratively, from the highest to the lowest,
until it reaches the first two indices. When this occurs, the function
calls sub2ind2d again.

- The "ArrayOfDimensions" is an input that holds the number of elements
in the two dimensions of this array. The "arraySize" is an input that
represents the number of dimensions of our array.
- The "index" input represents the linear index to be converted.
- And the "convertedIndex" is an array containing both subscripts.
The first is the row, and the second is the column.

See the demo/demo_3DArray.c file for more details.