feat: Add the sub2ind2D function
This new feature reverses the ind2sub2D function. That is, the
subscription is requested, and the sub2ind2D function spits the linear
index.

- The "arrayOfDimensions" is an input that holds the number of elements
in the two dimensions of this array.
- And the "arrayOfSubscripts" is an input array, containing both
subscripts. The first is the row, and the second is the column.

See the demo/demo_sub2ind2D.c file for more details.