feat: feat: Add the sub2ind function
As the new feature of the previous commit, this sub2ind function
converts subscripts to linear index, but for dimensions greater than 2.

- The "arrayOfDimensions" is an input that holds the number of
elements in the two dimensions of this array.
- The "arraySize" is an input that represents the size of the
"arrayOfDimensions".
- And the "arrayOfSubscripts" is an input array, containing both
subscripts. The first is the row, and the second is the column.

See the demo/demo_sub2ind.c file for more details.